Update to PVR UM DDK 1.12@2619570 for Android-4.4/Kit-Kat.
This is the same UM DDK code we used in Android-4.3;
note however that the binaries are still different.

Currently working on a new 1.12 change set which will
have a fix for for the Chrome problem and expect to
have that available shortly here and with a new
kernel PVR DDK update.

Built from current DDK-UM source:
    https:mipsia.review.mips.com:29418/local/pvrum

With:
   Branch:    klp-pdk
   Commit:    8f8efa3d84c81b98bc84061873e08a5a5ecf000e
   Change-Id: I7f84178a45040230f28560718e526c3e54ee2434

Easily installed here by:
    cd ${ANDROID_BUILD_TOP}/hardware/ingenic/DDK-UM
    make
    make install

Change-Id: Ica4eebdced815bfcd6bda668ab23d3da9f1fd512
Signed-off-by: Pete Delaney <piet.delaney@imgtec.com>