Remove Ingenic hwcomposer-SGX540 directory.
We've found that the Ingenic value-add in this module is not
actually working in the current MIPS tree. As a result the module
does nothing but force composition to be done by the GPU, which is
what would happen if there was no hwcomposer module anyway.

Since we are about to replace this hwcomposer with the stock IMG
implementation, remove it.

Change-Id: Iec69106125a24ab9c15a8f832beea6a3a9e9e6ba
Signed-off-by: Alistair Strachan <alistair.strachan@imgtec.com>