Update Camera HAL and XBOMX to use SGX function to provide Stride
Providing a new function, iStride(), which is defined in iStride.h,
that uses the Imagination pfnComputeParams() function to provide the
Horizontal Stride.

This is preferable to the previous modification by Ingenic
that adds iStride to the GPU buffer handle. This PVR approach is
needed by other PowerVR users and facilitates framebuffers laid out
for better cache line efficiency.

This commit is based on the following change:
"Update XBOMX to use SGX function to provide Stride."
by Pete Delaney <piet.delaney@imgtec.com>

Change-Id: I3542a4ace43fa05358e58849422b9644ad9be898
Signed-off-by: Dragan Cecavac <dragan.cecavac@imgtec.com>