Avoid audio driver logspam
It's considered normal for get_presentation_position() function
to fail sometimes. In those cases we return -EINVAL so that the error log
gets filtered. This way we avoid spamming the logcat.
In case of detecting underrun return 0 bytes written, with the same purpose.

This patch fixes the following repeated errors:
06-13 04:00:49.113  1059  1928 W StreamHAL: Error from HAL stream in function get_presentation_position: Operation not permitted
06-13 04:00:49.114  1059  1928 W StreamHAL: Error from HAL stream in function get_presentation_position: Operation not permitted
06-13 04:00:49.183  1059  1928 W StreamHAL: Error from HAL stream in function write: Broken pipe
06-13 04:00:49.183  1059  1103 E AudioStreamOutSink: Error while writing data to HAL: -61

Change-Id: I133771e0c882b6c631da16b11b3cf025d6489458
Signed-off-by: Dragan Cecavac <dragan.cecavac@imgtec.com>